cap: move the fp8_kv_scale_count tests to weights/fp8_scale_count_tests.rs
weights.rs landed at 506 with the tests inline; the sibling file puts it
at 470 with headroom. Tests self-contained (own store construction).
