+++ crates/spark-runtime/src/weights.rs
@@ -292,6 +292,9 @@ pub(crate) use loader::estimate_load_bytes;
 pub(crate) use loader::{check_oom_guard, estimate_has_fp8};
 
 #[cfg(test)]
+mod fp8_scale_count_tests;
+
+#[cfg(test)]
 mod from_str_tests {
     use super::WeightDtype;
 
@@ -449,45 +452,6 @@ mod teardown_tests {
         store.release(&gpu).expect("first");
         store.release(&gpu).expect("second");
         assert_eq!(gpu.alloc_count(), 0);
-    }
-
-    /// `fp8_kv_scale_count` counts exactly the `*.k_scale` tensors — one per
-    /// attention layer in checkpoints that ship calibrated FP8 KV scales —
-    /// and ignores `v_scale` (paired 1:1 with `k_scale`, counting both would
-    /// double-report) and lookalike suffixes.
-    #[test]
-    fn fp8_kv_scale_count_counts_only_k_scale_tensors() {
-        let gpu = MockGpuBackend::new();
-        let tensor = || WeightTensor {
-            ptr: gpu.alloc(1024).expect("alloc"),
-            shape: vec![1],
-            dtype: WeightDtype::BF16,
-        };
-        let mut map = HashMap::new();
-        for name in [
-            "model.layers.0.self_attn.k_scale",
-            "model.layers.0.self_attn.v_scale",
-            "model.layers.7.self_attn.k_scale",
-            "model.layers.7.self_attn.v_scale",
-            "model.layers.0.self_attn.q_proj.weight",
-            // Lookalikes that must NOT count: no dot before the suffix, and a
-            // different scale kind entirely.
-            "model.layers.0.self_attn.attnk_scale",
-            "model.layers.0.mlp.weight_scale",
-        ] {
-            map.insert(name.to_string(), tensor());
-        }
-        let store = WeightStore::from_map(map);
-        assert_eq!(store.fp8_kv_scale_count(), 2);
-    }
-
-    /// A checkpoint without shipped KV scales reports zero — the case where
-    /// serve logs the "needs calibration or a non-FP8 KV dtype" warning.
-    #[test]
-    fn fp8_kv_scale_count_zero_without_scales() {
-        let gpu = MockGpuBackend::new();
-        let store = store_with(&gpu, 4);
-        assert_eq!(store.fp8_kv_scale_count(), 0);
     }
 
     /// Reverse order, and one failure does not abandon the rest — the whole
